% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L.R
\name{L}
\alias{L}
\title{lkj
Return correlation matrix from random parameter}
\usage{
L(k, y)
}
\arguments{
\item{k}{integer. The dimension of the matrix}

\item{y}{The number of parameters to map.
The number of parameters should be is k*(k-1)/2}
}
\value{
A Square correlation matrix
}
\description{
lkj
Return correlation matrix from random parameter
}
\examples{
L(2, 1)
}
