% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_matrix.R
\name{L_matrix}
\alias{L_matrix}
\title{lkj
Return the Cholesky factor of the correlation matrix from random parameter}
\usage{
L_matrix(y)
}
\arguments{
\item{y}{The number of parameters to map.
The number of parameters should be is k*(k-1)/2}
}
\value{
A Square correlation matrix
}
\description{
lkj
Return the Cholesky factor of the correlation matrix from random parameter
}
\examples{
L_matrix(0.1)
}
