% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_par.R
\name{L_par}
\alias{L_par}
\title{lkj
Return the parameters  of a Cholesky factor of the correlation matrix}
\usage{
L_par(L)
}
\arguments{
\item{L}{Matrix. The Cholesky factor matrix
The number of parameters should be is k*(k-1)/2}
}
\value{
A vector of parameters of k*(k-1)/2 lenght
}
\description{
lkj
Return the parameters  of a Cholesky factor of the correlation matrix
}
\examples{
L <- L_matrix(0.1)
L_par(L)
}
