% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_matrix.R
\name{R_matrix}
\alias{R_matrix}
\title{lkj
Return correlation matrix from random parameter}
\usage{
R_matrix(k, y)
}
\arguments{
\item{k}{integer. The dimension of the matrix}

\item{y}{The number of parameters to map.
The number of parameters should be is k*(k-1)/2}
}
\value{
A Square correlation matrix
}
\description{
lkj
Return correlation matrix from random parameter
}
\examples{
R_matrix(2, 0.1)
}
