% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ronion.R
\name{ronion}
\alias{ronion}
\title{Generate multiple correlation matrices via the onion method}
\usage{
ronion(n, d, eta = 1)
}
\arguments{
\item{n}{Number of matrices to generate.}

\item{d}{Dimension of the matrices.}

\item{eta}{Positive shape parameter.}
}
\value{
If \code{n = 1}, a \code{d x d} correlation matrix.
  If \code{n > 1}, an array of dimension \code{c(d, d, n)}.
}
\description{
Generate multiple correlation matrices via the onion method
}
\examples{
R_list <- ronion(5, d = 3, eta = 1)

}
